access control

+++ app/Http/Controllers/PostsController.php
@@ -9,6 +9,16 @@ use App\Models\BlogPost;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -76,6 +86,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = BlogPost::find($id);
+
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('posts')->with('error', 'Unauthorized Page');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -110,6 +125,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = BlogPost::find($id);
+
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('posts')->with('error', 'Unauthorized Page');
+        }
+
         $post->delete();
         return redirect('posts')->with('success', 'Post Deleted');
     }
